\name{ctrldefault}
\alias{ctrldefault}
\encoding{UTF-8}
\title{Set default values of control parameters for routines in the \code{optimz} package}
\description{Set default values of control parameters for routines in the \code{optimz}
    package. See the source code for the names and values. 
}
\usage{
       ctrldefault(npar)
}
\arguments{
 \item{npar}{The (scalar) number of parameters in the function to be optimized.}
}

\details{
None.
}

\value{
  A list with components as described in the source code.

}
\examples{
cat("all examples in tests/ subdirectory\n")
}
\keyword{nonlinear}
\keyword{optimize}

